Bugfixes and 64bit compatibilty patch that was submitted by Christoph
--HG--
extra : convert_revision : svn%3A0a3f77c4-9235-0410-95f4-d138c4647c51/trunk%40771

+++ IMBNodeCell.m
@@ -248,7 +248,7 @@
 //----------------------------------------------------------------------------------------------------------------------
 
 
-- (void) selectWithFrame:(NSRect)inFrame inView:(NSView*)inControlView editor:(NSText*)inText delegate:(id)inDelegate start:(int)inStart length:(int)inLength
+- (void) selectWithFrame:(NSRect)inFrame inView:(NSView*)inControlView editor:(NSText*)inText delegate:(id)inDelegate start:(NSInteger)inStart length:(NSInteger)inLength
 {
 	NSRect titleRect = [self titleRectForBounds:inFrame];
 	[super selectWithFrame:titleRect inView:inControlView editor:inText delegate:inDelegate start:inStart length:inLength];
